Separate mleap tests
